# Single root fast path