--- a/archiver_test.go
+++ b/archiver_test.go
@@ -14,19 +14,20 @@ import (
 	"github.com/src-d/borges/lock"
 	"github.com/src-d/borges/storage"
 
-	"github.com/satori/go.uuid"
+	uuid "github.com/satori/go.uuid"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"gopkg.in/src-d/core-retrieval.v0/model"
 	"gopkg.in/src-d/core-retrieval.v0/repository"
 	"gopkg.in/src-d/core-retrieval.v0/test"
-	"gopkg.in/src-d/go-billy.v4"
+	billy "gopkg.in/src-d/go-billy.v4"
 	"gopkg.in/src-d/go-billy.v4/osfs"
-	"gopkg.in/src-d/go-git-fixtures.v3"
-	"gopkg.in/src-d/go-git.v4"
+	fixtures "gopkg.in/src-d/go-git-fixtures.v3"
+	git "gopkg.in/src-d/go-git.v4"
+	"gopkg.in/src-d/go-git.v4/plumbing"
 	"gopkg.in/src-d/go-git.v4/plumbing/object"
 	"gopkg.in/src-d/go-git.v4/storage/memory"
-	"gopkg.in/src-d/go-kallax.v1"
+	kallax "gopkg.in/src-d/go-kallax.v1"
 )
 
 func TestArchiver(t *testing.T) {
@@ -41,6 +42,7 @@ type ArchiverSuite struct {
 	store    RepositoryStore
 	tmpPath  string
 	tx       repository.RootedTransactioner
+	copier   *repository.Copier
 	txFs     billy.Filesystem
 	tmpFs    billy.Filesystem
 	rootedFs billy.Filesystem
@@ -71,18 +73,19 @@ func (s *ArchiverSuite) SetupTest() {
 	s.tmpFs, err = fs.Chroot("tmp")
 	s.NoError(err)
 
-	copier := repository.NewCopier(
+	s.copier = repository.NewCopier(
 		s.txFs,
 		repository.NewLocalFs(s.rootedFs),
 		s.bucket)
-	s.tx = repository.NewSivaRootedTransactioner(copier)
+	s.tx = repository.NewSivaRootedTransactioner(s.copier)
 
 	ls, err := lock.NewLocal().NewSession(&lock.SessionConfig{
 		Timeout: defaultTimeout,
 	})
 	s.NoError(err)
 
-	s.a = NewArchiver(s.store, s.tx, NewTemporaryCloner(s.tmpFs), ls, defaultTimeout)
+	s.a = NewArchiver(s.store, s.tx, NewTemporaryCloner(s.tmpFs),
+		ls, defaultTimeout, s.copier)
 }
 
 func (s *ArchiverSuite) TearDownTest() {
@@ -147,7 +150,8 @@ func (s *ArchiverSuite) TestLockTimeout() {
 	start := time.Now()
 
 	job := &Job{RepositoryID: uuid.UUID(repoUUID)}
-	a := NewArchiver(s.store, s.tx, NewTemporaryCloner(s.tmpFs), session, 10*time.Second)
+	a := NewArchiver(s.store, s.tx, NewTemporaryCloner(s.tmpFs),
+		session, 10*time.Second, s.copier)
 
 	ctx := context.TODO()
 	err = a.Do(ctx, job)
@@ -174,7 +178,7 @@ func (s *ArchiverSuite) TestReferenceUpdate() {
 }
 
 func (s *ArchiverSuite) getFileNames(p string) ([]string, error) {
-	files := make([]string, 10)
+	var files []string
 
 	dirents, err := s.rootedFs.ReadDir(p)
 	if err != nil {
@@ -243,21 +247,67 @@ func (s *ArchiverSuite) TestFixtures() {
 			}
 			checkReferencesInDB(t, mr, ct.NewReferences)
 
-			// check references in siva files
+			type references map[plumbing.ReferenceName]bool
+
 			fis, err := s.getFileNames(".")
 			if len(ct.NewReferences) != 0 {
 				require.NoError(err)
-				initHashesInStorage := make(map[string]bool)
+				initHashesInStorage := make(map[string]references)
 
 				for _, fi := range fis {
-					initHashesInStorage[strings.Replace(fi, ".siva", "", -1)] = true
+					hashStr := strings.Replace(fi, ".siva", "", -1)
+					hash := plumbing.NewHash(hashStr)
+
+					// get siva file and open it
+					tx, err := s.a.RootedTransactioner.Begin(context.TODO(), hash)
+					require.NoError(err)
+
+					repo, err := git.Open(tx.Storer(), nil)
+					require.NoError(err)
+
+					// gather references for later check
+					it, err := repo.References()
+					require.NoError(err)
+
+					refs := make(references)
+					err = it.ForEach(func(r *plumbing.Reference) error {
+						refs[r.Name()] = true
+						return nil
+					})
+					require.NoError(err)
+
+					initHashesInStorage[hashStr] = refs
+
+					remotes, err := repo.Remotes()
+					require.NoError(err)
+
+					// check that "origin" remote does not exist and that
+					// the one downloaded is configured
+					var found bool
+					for _, r := range remotes {
+						require.NotEqual("origin", r.Config().Name)
+						if r.Config().Name == rid.String() {
+							found = true
+						}
+					}
+					require.True(found)
+
+					// delete previously copied siva file
+					err = tx.Rollback()
+					require.NoError(err)
 				}
 
-				// we check that all the references that we have into the database exists as a rooted repository
+				// check that all the references that we have into the
+				// database exists as a rooted repository
 				for _, ref := range mr.References {
-					_, ok := initHashesInStorage[ref.Init.String()]
+					r, ok := initHashesInStorage[ref.Init.String()]
+					require.True(ok)
+
+					name := rootedRefName(ref.GitReference().Name(), mr.ID)
+					_, ok = r[name]
 					require.True(ok)
 				}
+
 			}
 		})
 	}
@@ -423,7 +473,8 @@ func customArchiver(
 	})
 	require.NoError(err)
 
-	a := NewArchiver(store, tx, NewTemporaryCloner(tmpFs), ls, defaultTimeout)
+	a := NewArchiver(store, tx, NewTemporaryCloner(tmpFs),
+		ls, defaultTimeout, copier)
 
 	repoFixture := fixtures.ByTag("worktree").One()
 	repoPath := repoFixture.Worktree().Root()
@@ -588,8 +639,8 @@ type BrokenFile struct {
 }
 
 func (fs *BrokenFile) Write(p []byte) (int, error) {
-	// A siva copy takes less than 10 writes to complete, it should be a push.
-	if fs.count > 10 {
+	// A siva copy takes less than 2 writes to complete, it should be a push.
+	if fs.count > 2 {
 		return 0, fmt.Errorf("could not write to broken file")
 	}